refactor: implement snapshotCurrentGrimoire function in history module and update its usage
