+++ ui/history/grimoire.js
@@ -1,3 +1,6 @@
+import { generateId, formatDateName } from "../../utils.js";
+import { saveHistories } from "./index.js";
+
 export function renderGrimoireHistory({ grimoireHistoryList, history }) {
   if (!grimoireHistoryList) return;
   grimoireHistoryList.innerHTML = '';
@@ -35,4 +38,23 @@ export function renderGrimoireHistory({ grimoireHistoryList, history }) {
     li.appendChild(deleteBtn);
     grimoireHistoryList.appendChild(li);
   });
+}
+
+export function snapshotCurrentGrimoire({ players, scriptMetaName, scriptData, history, grimoireHistoryList }) {
+  try {
+    if (!Array.isArray(players) || players.length === 0) return;
+    const snapPlayers = JSON.parse(JSON.stringify(players));
+    let name = formatDateName(new Date());
+    const entry = {
+      id: generateId('grimoire'),
+      name,
+      createdAt: Date.now(),
+      players: snapPlayers,
+      scriptName: scriptMetaName || (Array.isArray(scriptData) && (scriptData.find(x => x && typeof x === 'object' && x.id === '_meta')?.name || '')) || '',
+      scriptData: Array.isArray(scriptData) ? JSON.parse(JSON.stringify(scriptData)) : null
+    };
+    history.grimoireHistory.unshift(entry);
+    saveHistories(history);
+    renderGrimoireHistory({ grimoireHistoryList, history });
+  } catch (_) { }
 }
